Use hand_detection.h
